Move top level xml importer to separate file
